Added Scaffold Ray debug options
[ChestAura & Fucker]
* Improved smooth range easing animation

[Scaffold]
+ Added Visual.Debug.Ray options

+++ shared/main/java/net/ccbluex/liquidbounce/ui/client/clickgui/style/styles/SlowlyStyle.kt
@@ -64,7 +64,7 @@ class SlowlyStyle : Style()
 		val font = getDescriptionFont()
 		val fontHeight = font.fontHeight
 		val textWidth = font.getStringWidth(text)
-		drawBorderedRect((mouseX + 9).toFloat(), mouseY.toFloat(), (mouseX + textWidth + 14).toFloat(), (mouseY + fontHeight + 3).toFloat(), 3.0f, BACKGROUND, BACKGROUND)
+		drawBorderedRect((mouseX + 9).toFloat(), mouseY.toFloat(), mouseX + textWidth + 14f, mouseY + fontHeight + 3f, 3.0f, BACKGROUND, BACKGROUND)
 
 		classProvider.glStateManager.resetColor()
